// src/types/generated-admin-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Adjustment = {
  __typename?: 'Adjustment';
  adjustmentSource: Scalars['String'];
  type: AdjustmentType;
  description: Scalars['String'];
  amount: Scalars['Int'];
};

export enum AdjustmentType {
  Promotion = 'PROMOTION',
  DistributedOrderPromotion = 'DISTRIBUTED_ORDER_PROMOTION',
  Other = 'OTHER'
}
// ...
export enum NfeEnv {
  Production = 'PRODUCTION',
  Homologation = 'HOMOLOGATION'
}
// ...
export type NfeOperationInfo = {
  __typename?: 'NfeOperationInfo';
  classe_imposto?: Maybe<Scalars['String']>;
  operacao?: Maybe<Scalars['Int']>;
  natureza_operacao?: Maybe<Scalars['String']>;
  volume?: Maybe<Scalars['Float']>;
  informacoes_fisco?: Maybe<Scalars['String']>;
  informacoes_complementares?: Maybe<Scalars['String']>;
  correcao?: Maybe<Scalars['String']>;
  motivo?: Maybe<Scalars['String']>;
};

export type NfeOperationProduct = {
  __typename?: 'NfeOperationProduct';
  quantidade?: Maybe<Scalars['Float']>;
  subtotal?: Maybe<Scalars['Float']>;
  total?: Maybe<Scalars['Float']>;
  codigo_cfop?: Maybe<Scalars['String']>;
  situacao_tributaria?: Maybe<Scalars['String']>;
  beneficio_fiscal?: Maybe<Scalars['String']>;
};

export type NfeOperationResponse = {
  __typename?: 'NfeOperationResponse';
  emissorId?: Maybe<Scalars['String']>;
  status?: Maybe<Scalars['String']>;
  serie?: Maybe<Scalars['String']>;
  numero?: Maybe<Scalars['String']>;
  recibo?: Maybe<Scalars['String']>;
  chave?: Maybe<Scalars['String']>;
  xml?: Maybe<Scalars['String']>;
  danfe?: Maybe<Scalars['String']>;
  sefaz?: Maybe<Scalars['String']>;
};

export type NfeOperationTaxes = {
  __typename?: 'NfeOperationTaxes';
  codigo_cfop?: Maybe<Scalars['String']>;
  beneficio_fiscal?: Maybe<Scalars['String']>;
  situacao_tributaria?: Maybe<Scalars['String']>;
  item_servico?: Maybe<Scalars['Int']>;
  exigibilidade?: Maybe<Scalars['Int']>;
  incentivo_fiscal?: Maybe<Scalars['Int']>;
  bc_ipi?: Maybe<Scalars['Float']>;
  bc_icms_st?: Maybe<Scalars['Float']>;
  bc_issqn?: Maybe<Scalars['Float']>;
  valor_ipi?: Maybe<Scalars['Float']>;
  valor_icms?: Maybe<Scalars['Float']>;
  valor_icms_st?: Maybe<Scalars['Float']>;
  valor_issqn?: Maybe<Scalars['Float']>;
  aliquota_mva?: Maybe<Scalars['Float']>;
};

export enum NfeOperationType {
  Emission = 'EMISSION',
  Devolution = 'DEVOLUTION',
  Adjustment = 'ADJUSTMENT',
  Complementary = 'COMPLEMENTARY',
  Correction = 'CORRECTION',
  Cancel = 'CANCEL'
}